# perf : 검사 결과를 통계 정보로 변환하는 Batch 성능 개선 