--- a/batch/src/main/resources/application.yml
+++ b/batch/src/main/resources/application.yml
@@ -8,7 +8,7 @@ spring:
 
   batch:
     job:
-      name: statisticJob
+      name: statisticJobNew
       enabled: true
     jdbc:
       initialize-schema: always