--- a/domain/src/main/java/org/meotppo/webti/domain/config/mongo/MongoConfig.java
+++ b/domain/src/main/java/org/meotppo/webti/domain/config/mongo/MongoConfig.java
@@ -2,7 +2,10 @@ package org.meotppo.webti.domain.config.mongo;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.MongoDatabaseFactory;
+import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.convert.MongoCustomConversions;
+
 import java.util.List;
 
 @Configuration
@@ -13,5 +16,10 @@ public class MongoConfig {
             LocalDateTimeToDateKstConverter localDateTimeToDateKstConverter,
             DateToLocalDateTimeKstConverter dateToLocalDateTimeKstConverter) {
         return new MongoCustomConversions(List.of(localDateTimeToDateKstConverter, dateToLocalDateTimeKstConverter));
+    }
+
+    @Bean
+    public MongoTemplate mongoTemplate(MongoDatabaseFactory mongoDbFactory) {
+        return new MongoTemplate(mongoDbFactory);
     }
 }