--- a/domain/src/main/java/org/meotppo/webti/domain/config/mongo/MongoConfig.java
+++ b/domain/src/main/java/org/meotppo/webti/domain/config/mongo/MongoConfig.java
@@ -4,7 +4,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.MongoDatabaseFactory;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.convert.DefaultMongoTypeMapper;
+import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
 import org.springframework.data.mongodb.core.convert.MongoCustomConversions;
+import org.springframework.data.mongodb.core.mapping.MongoMappingContext;
 
 import java.util.List;
 
@@ -19,7 +22,11 @@ public class MongoConfig {
     }
 
     @Bean
-    public MongoTemplate mongoTemplate(MongoDatabaseFactory mongoDbFactory) {
-        return new MongoTemplate(mongoDbFactory);
+    public MongoTemplate mongoTemplate(MongoDatabaseFactory mongoDbFactory, MongoMappingContext context, MongoCustomConversions customConversions) {
+        MappingMongoConverter converter = new MappingMongoConverter(mongoDbFactory, context);
+        converter.setTypeMapper(new DefaultMongoTypeMapper(null));
+        converter.setCustomConversions(customConversions);
+        converter.afterPropertiesSet();
+        return new MongoTemplate(mongoDbFactory, converter);
     }
 }