<!--
PR 이름 컨벤션
feat: ~~(#issueNum)
-->

## 📌 관련 이슈

- closed: #37

## ✨ PR 세부 내용
- 기존의 reader, processor, writer 구조를 단일 tasklet으로 재구성
- MongoDB Aggregation을 이용하여 데이터 변환 및 통계 처리
- 기존 코드와 성능 비교를 통해 개선 여부 검증

## 📄 개선 결과
#### 개선 전
- **16 * 10 테스트 (16개의 MBTI 유형 각각에 대해 10개의 테스트 데이터 처리)**
  - 1차: 26s 971ms
  - 2차: 27s 102ms
  - 3차: 26s 843ms

- **16 * 50 테스트 (16개의 MBTI 유형 각각에 대해 50개의 테스트 데이터 처리)**
  - 1차: 2m 4s 892ms
  - 2차: 2m 5s 312ms
  - 3차: 2m 4s 756ms

#### 개선 후
- **16 * 10 테스트 (16개의 MBTI 유형 각각에 대해 10개의 테스트 데이터 처리)**
  - 1차: 5s 332ms
  - 2차: 5s 213ms
  - 3차: 5s 348ms

- **16 * 50 테스트 (16개의 MBTI 유형 각각에 대해 50개의 테스트 데이터 처리)**
  - 1차: 5s 357ms
  - 2차: 5s 298ms
  - 3차: 5s 376ms

기존 구조에서 단일 Tasklet으로 재구성하고 MongoDB Aggregation을 활용한 결과, 성능이 크게 개선되었습니다. 또한, 데이터의 양이 증가해도 처리 시간이 거의 일정하게 유지됨을 확인했습니다.