@DongHyeonka 2024-07-03
네 구체적인 답변 감사합니다 다음에 공부할 때 참고해야겠네요